Refactor: Split config_flow.py and coordinator.py into smaller modules
Split the two largest files for better maintainability:

config_flow.py (1,550 → ~1,100 lines):
- Extract detection.py: bed type detection logic (~280 lines)
- Extract validators.py: validation and config helpers (~90 lines)

coordinator.py (1,625 → ~1,100 lines):
- Extract adapter.py: Bluetooth adapter selection and BLE helpers (~270 lines)
- Extract controller_factory.py: controller instantiation logic (~200 lines)

This is a pure refactoring with no functional changes. All 494 tests pass.

Closes #98
